+++ pytsite/auth/res/assets/js/auth.js
@@ -1,12 +1,12 @@
 pytsite.auth = {
     getLoginForm: function (driver, title, uid, css, modal, success) {
-        pytsite.jsApi.get('pytsite.auth.get_login_form', {
+        pytsite.httpApi.get('pytsite.auth.get_login_form', {
             driver: driver,
             title: title,
             uid: uid,
             css: css,
             modal: modal
-        }, function (resp) {
+        }).done(function (resp) {
             if (success)
                 success(resp);
         });